@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
+import { OKTA_AUTH } from '@okta/okta-angular';
 import { OktaAuth } from '@okta/okta-auth-js';
 
 @Component({
@@ -10,13 +11,10 @@ export class AccountInfoComponent implements OnInit {
   userName!: string;
   userEmail!: string;
 
-  constructor(private oktaAuth: OktaAuth) {
-
-  }
+  constructor(@Inject(OKTA_AUTH) private oktaAuth: OktaAuth) {}
 
   ngOnInit(): void {
-    this.oktaAuth.getUser().then(
-      user => {
+    this.oktaAuth.getUser().then(user => {
         this.userName = user.name as string;
         this.userEmail = user.email as string;
       }
